add tests for multirequests
